+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [unreleased]
 ### Changed
 - Log configured sensors
+- Round sensor values
 
 ## [1.0.2] - 2023-08-22
 ### Changed